--- src/main/java/com/shootdoori/match/dto/MercenaryReviewResponseDto.java
@@ -1,4 +1,0 @@
-package com.shootdoori.match.dto;
-
-public record MercenaryReviewResponseDto() {
-}
